remove puts and gets into main.rb
