+++ lib/board.rb
@@ -9,7 +9,7 @@ class Board
     @board[pos[0]][pos[1]] = type
   end
 
-  # Check if user input is valid
+ 
   def valid_input?(pos)
     arr = pos.split('')
     x_axis = arr[1]
@@ -25,15 +25,7 @@ class Board
     false
   end
 
-  def status
-    if check_if_x_win || check_if_y_win || check_if_diag_win
-      puts 'win'
-      1
-    else
-      puts 'no win'
-    end
- end
-
+  
   def winner(player)
     b = @board
 
@@ -50,33 +42,16 @@ class Board
   end
 
   def draw
-    puts '    ___________'
-    puts "C  |_#{@board[2][0]}_|_#{@board[2][1]}_|_#{@board[2][2]}_|"
-    puts "B  |_#{@board[1][0]}_|_#{@board[1][1]}_|_#{@board[1][2]}_|"
-    puts "A  |_#{@board[0][0]}_|_#{@board[0][1]}_|_#{@board[0][2]}_|"
-    puts "     1   2   3\t\t"
+    output = ""
+     output += "\n __________________"
+
+     output += "\nC  |_#{@board[2][0]}_|_#{@board[2][1]}_|_#{@board[2][2]}_|"
+     output += "\nB  |_#{@board[1][0]}_|_#{@board[1][1]}_|_#{@board[1][2]}_|"
+     output += "\nA  |_#{@board[0][0]}_|_#{@board[0][1]}_|_#{@board[0][2]}_|"
+     output += "\n     1   2   3\t\t"
+     output
   end
 
-  private
 
-  def check_if_y_win
-    @board.each do |y_axis|
-      return y_axis.all?('X') || y_axis.all?('O')
-    end
-  end
-
-  def check_if_x_win
-    @board.each do |x_axis|
-      return x_axis.all?('X') || x_axis.all?('O')
-    end
-  end
-
-  def check_if_diag_win
-    if @board[1][1] != '_'
-      diag1 = @board[0][0] == @board[1][1] && @board[1][1] == @board[2][2]
-      diag2 = @board[0][2] == @board[1][1] && @board[1][1] == @board[2][0]
-      return diag1 || diag2
-    end
-    false
-  end
+  
 end
